[fix] divide map component
