+++ front/src/pages/MapPage.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react'
 import { Container, Row, Col, ListGroup, ListGroupItem } from 'react-bootstrap'
-import { Chart } from "react-google-charts"
-import Item from './ListItem'
-const googleAPIKey = "AIzaSyBTvsuJcbhSf2giulYdP66791797JE4ZTA"
+import Item from '../components/MapPage/ListItem'
+import Map from '../components/MapPage/Map'
 
 /*
 입국 금지 : 검정 (0)
@@ -10,17 +9,15 @@ const googleAPIKey = "AIzaSyBTvsuJcbhSf2giulYdP66791797JE4ZTA"
 입국 제한 : 빨강 (1)
   - 국가명, 입국제한조치 제목 출력, 기준일자시간, 국가명 옆의 칸인 입국 제한 조치
 해당 없음 : default
+
+data 형식
+    -> ["나라명(영어)", "상태", "디테일(툴팁용)"]
 */
 
 let items = []
 let data = [
   ['Country', 'State', { role: "tooltip", type: "string", p: { html: true } }]
 ]
-
-const options = {
-  colorAxis: { colors: ['black', 'red'] },
-  tooltip: { isHtml: true, trigger: "visible" }
-};
 
 class MapPage extends Component {
 
@@ -29,10 +26,7 @@ class MapPage extends Component {
     lists: items // 리스트뷰 데이터
   }
 
-  /*
-  data 형식
-    -> ["나라명(영어)", "상태", "디테일(툴팁용)"]
-  */
+ 
   getRestrictionData = async () => {
     const response = await fetch('/map')
     const body = await response.json()
@@ -63,22 +57,17 @@ class MapPage extends Component {
 
   render() {
     return <div>
+      
       <Container>
-        <Chart
-          width={'100%'}
-          height={'100%'}
-          chartType="GeoChart"
-          mapsApiKey={googleAPIKey}
-          data={this.state.countries}
-          options={options}
-        />
+        <Map countries = {this.state.countries}/>
       </Container>
 
       <Container>
         <ListGroup>
-          <Item items={this.state.lists}/>
+          <Item lists={this.state.lists}/>
         </ListGroup>
       </Container>
+    
     </div>
   }
 }
